+++ src/main/java/lotto/validator/InputValidator/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto.validator;
+package lotto.validator.InputValidator;
 
 public interface InputValidator<T> {
     String ERROR_MESSAGE = "[ERROR]";
